# apps/api/internal/platform/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Orthodontics Helper API
  version: 0.1.0
  description: Clinic-scoped patient, appointment, SMS, and audit API.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      responses:
        "200":
          description: API health status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/auth/me:
    get:
      operationId: getCurrentUser
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Error"
  /v1/auth/profile:
    patch:
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/auth/password:
    post:
      operationId: updatePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordInput"
      responses:
        "200":
          description: Password updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/clinics/current:
    get:
      operationId: getCurrentClinic
      responses:
        "200":
          description: Current clinic.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          $ref: "#/components/responses/Error"
  /v1/colleagues:
    get:
      operationId: listColleagues
      responses:
        "200":
          description: Clinic colleagues.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColleagueListResponse"
  /v1/clinic-members:
    post:
      operationId: createClinicMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClinicMemberInput"
      responses:
        "201":
          description: Created clinic member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Colleague"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/patients:
    get:
      operationId: listPatients
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Clinic patients.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created patient.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/Error"
  /v1/patients/{id}:
    get:
      operationId: getPatient
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Patient detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updatePatient
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "200":
          description: Updated patient.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Patient soft deleted.
        "404":
          $ref: "#/components/responses/Error"
  /v1/appointments:
    get:
      operationId: listAppointments
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
      responses:
        "200":
          description: Clinic appointments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /v1/appointments/{id}:
    patch:
      operationId: updateAppointment
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "200":
          description: Updated appointment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Appointment soft deleted.
  /v1/sms/templates:
    get:
      operationId: listSMSTemplates
      responses:
        "200":
          description: Clinic SMS templates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSTemplateListResponse"
  /v1/sms/send:
    post:
      operationId: sendManualSMS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendManualSMSInput"
      responses:
        "201":
          description: SMS send result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendSMSResponse"
        "502":
          $ref: "#/components/responses/Error"
  /v1/audit-logs:
    get:
      operationId: listAuditLogs
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AuditEntityType"
        - name: entityId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entity audit logs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Stable error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
    CurrentUser:
      type: object
      required: [id, email, fullName, clinicId]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        clinicId:
          type: string
          format: uuid
    UpdateProfileInput:
      type: object
      required: [fullName, email]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
    UpdatePasswordInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
    Clinic:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Colleague:
      type: object
      required: [id, fullName, email, clinic, joinedAt]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        clinic:
          type: string
        joinedAt:
          type: string
          format: date-time
    ColleagueListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Colleague"
    CreateClinicMemberInput:
      type: object
      required: [fullName, email, temporaryPassword]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        temporaryPassword:
          type: string
          minLength: 8
    PatientStatus:
      type: string
      enum: [active_treatment, completed, cancelled, waiting, inactive]
    Patient:
      type: object
      required: [id, fullName, phone, status, remindersEnabled, reminderDaysBefore, reminderSendHour, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/PatientStatus"
        treatmentNote:
          type: string
        internalNote:
          type: string
        remindersEnabled:
          type: boolean
        reminderDaysBefore:
          type: array
          items:
            type: integer
        reminderSendHour:
          type: integer
          minimum: 0
          maximum: 23
        nextAppointment:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PatientInput:
      type: object
      required: [fullName, phone, status]
      properties:
        fullName:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/PatientStatus"
        treatmentNote:
          type: string
        internalNote:
          type: string
        remindersEnabled:
          type: boolean
        reminderDaysBefore:
          type: array
          items:
            type: integer
        reminderSendHour:
          type: integer
          minimum: 0
          maximum: 23
    PatientListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
    AppointmentStatus:
      type: string
      enum: [planned, confirmed, completed, cancelled, no_show]
    Appointment:
      type: object
      required: [id, patientId, startsAt, status]
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorUserId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        note:
          type: string
    AppointmentInput:
      type: object
      required: [patientId, startsAt, status]
      properties:
        patientId:
          type: string
          format: uuid
        doctorUserId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        note:
          type: string
    AppointmentListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    SMSTemplate:
      type: object
      required: [id, key, title, body, enabled]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        title:
          type: string
        body:
          type: string
        enabled:
          type: boolean
    SMSTemplateListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SMSTemplate"
    SendManualSMSInput:
      type: object
      required: [patientId, templateId]
      properties:
        patientId:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
    SendSMSResponse:
      type: object
      required: [messageId, status]
      properties:
        messageId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, sent, failed]
    AuditEntityType:
      type: string
      enum: [patient, appointment, sms_template, sms_message, notification_setting, clinic_member]
    AuditLog:
      type: object
      required: [id, action, summary, createdAt]
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"